requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
discord==2.3.2
discord.py==2.4.0
frozenlist==1.5.0
h11==0.14.0
idna==3.10
multidict==6.1.0
numpy==2.2.4
opencv-python==4.11.0.86
outcome==1.3.0.post0
packaging==24.2
pillow==11.2.1
propcache==0.2.0
PySocks==1.7.1
pytesseract==0.3.13
python-dotenv==1.0.1
redis==6.0.0
requests==2.32.3
selenium==4.25.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.18.0
